[MIG] pos_product_template: Migration to 17.0
