+++ pos_product_template/models/__init__.py
@@ -1,1 +1,2 @@
 from . import pos_config
+from . import res_config_settings
