--- pos_product_template/views/pos_config_view.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="pos_config_view_form" model="ir.ui.view">
-        <field name="name">pos.config.form.view</field>
-        <field name="model">pos.config</field>
-        <field name="inherit_id" ref="point_of_sale.pos_config_view_form" />
-        <field name="arch" type="xml">
-            <field name="fiscal_position_ids" position="after">
-                <field name="iface_product_template_show_variants" />
-            </field>
-        </field>
-    </record>
-</odoo>
