+++ pos_product_template/views/res_config_settings_view.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="res_config_settings_view_form" model="ir.ui.view">
+        <field name="name">res.config.settings.view.form</field>
+        <field name="model">res.config.settings</field>
+        <field name="inherit_id" ref="point_of_sale.res_config_settings_view_form" />
+        <field name="arch" type="xml">
+            <xpath expr="//setting[@id='margin_and_cost']" position="after">
+                <setting
+                    id="iface_product_template_show_variants"
+                    help="If selected the product variant selection screen will show the variants"
+                >
+                    <field name="iface_product_template_show_variants" />
+                </setting>
+            </xpath>
+        </field>
+    </record>
+</odoo>
